Test tenant and middleware for integration tests
Adds optional test_tenant and rack middleware to facilitate model
setup for integration tests. This addition seeks to solve a problem
during integration testing whereby data setup is done in a tenancy,
but the tenancy is not cleared prior to initiating a request due to
a lack of tolerance for the repetitive boilerplate that would be
required. The result is that integration testing is initiated with
a tenancy already set.