Add middleware option to reset tenancy
Adds optional rack middleware to the gem to allow the request-response
cycle reset the the current_tenant. This feature solves a problem
during integration testing whereby data setup is done in a tenancy,
but the tenancy is not dropped due to a lack of tolerance for the
repetitive boilerplate that would be required. The result is that
integration testing is initiated with a tenancy already set.